+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Modified
+
+- Push tag now commit and push the bump of version, then push tag
+
+### Fixed
+
+- Fixed the bug where the signup was not done for new users
+- Now on signup the message is better looking and well placed
+
 ## [1.4.0] - 2023-07-21
 
 Change ChatbotX name to AppifyAi